add therock-fftw3 and therock-fftw3f
Add for fftw3 builds configured with thread support for
single and double precision support.

FFTW (fftv3) will by default build the support for
double precision support and support for single
precision/float can not be enabled on same build.
Instead support for both precision types must be done by running
the configure and build phases for both of them separately.

fftw3 could also support a following precision types
that are not enabled in this PR:
- long double (extended-precision)
- quad double (quad double/128 bit) precision

Signed-off-by: Mika Laitio <mika.laitio@amd.com>
Co-authored-by: Geo Min <geomin12@amd.com>